+++ database/weather_information_collector.sql
@@ -38,6 +38,7 @@ CREATE TABLE IF NOT EXISTS `weatherdata` (
   `temperature_C` float DEFAULT NULL COMMENT 'temperature in degrees Celsius',
   `temperature_F` float DEFAULT NULL COMMENT 'temperature in degrees Fahrenheit',
   `humidity` tinyint(3) unsigned DEFAULT NULL COMMENT 'relative humidity in percent',
+  `rain` float DEFAULT NULL COMMENT 'amount of rain in millimeters',
   `pressure` smallint(5) unsigned DEFAULT NULL COMMENT 'air pressure in hPa',
   `wind_speed` float DEFAULT NULL COMMENT 'wind speed in meters per second',
   `wind_degrees` smallint(5) unsigned DEFAULT NULL COMMENT 'wind direction in degrees',